movie_recommender_project/requirements.txt
fastapi
uvicorn
pandas
scikit-surprise
matplotlib
seaborn
